Replace queue with linked list
Replace the bounded queue with a linked list and condvar
implementation, and replace the closed_slots system with double
indirection via AsyncClient's own memory.  This allows the
system to correctly handle cases where it is not possible to allocate a
new event and still ensure that the client's `onDisconnect()` will
be queued for execution.